+++ src/main/java/org/logicng/functions/SubNodeFunction.java
@@ -38,19 +38,14 @@ import java.util.LinkedHashSet;
 /**
  * A function that computes all sub-nodes of a given formula.  The order of the sub-nodes is bottom-up, i.e. a
  * sub-node only appears in the result when all of its sub-nodes are already listed.
- * @version 2.3.0
+ * @version 3.0.0
  * @since 1.0
  */
 public final class SubNodeFunction implements FormulaFunction<LinkedHashSet<Formula>> {
 
     private static final SubNodeFunction INSTANCE = new SubNodeFunction();
 
-    /**
-     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
-     * In order to instantiate an object of this class, use the {@link #get()} method.
-     */
-    @Deprecated
-    public SubNodeFunction() {
+    private SubNodeFunction() {
         // Intentionally left empty
     }