+++ src/main/java/org/logicng/functions/VariableProfileFunction.java
@@ -48,19 +48,14 @@ import java.util.TreeMap;
  * For this function, the non-caching version is preferred since it usually performs better.  The non-caching version
  * of this function generates the result mapping only once and fills it recursively whereas the caching version has to
  * construct a new mapping for each sub-formula.
- * @version 2.3.0
+ * @version 3.0.0
  * @since 1.0
  */
 public final class VariableProfileFunction implements FormulaFunction<Map<Variable, Integer>> {
 
     private static final VariableProfileFunction INSTANCE = new VariableProfileFunction();
 
-    /**
-     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
-     * In order to instantiate an object of this class, use the {@link #get()} method.
-     */
-    @Deprecated
-    public VariableProfileFunction() {
+    private VariableProfileFunction() {
         // Intentionally left empty
     }