+++ src/main/java/org/logicng/solvers/MaxSATSolver.java
@@ -58,7 +58,7 @@ import java.util.TreeSet;
 
 /**
  * A wrapper for the OpenWBO solver.
- * @version 2.4.0
+ * @version 3.0.0
  * @since 1.0
  */
 public class MaxSATSolver {
@@ -91,16 +91,6 @@ public class MaxSATSolver {
     }
 
     /**
-     * Returns a new MaxSAT solver using incremental WBO as algorithm with the default configuration.
-     * @return the MaxSAT solver
-     * @deprecated this factory method without parameter will be removed in the future, use {@link #incWBO(FormulaFactory)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver incWBO() {
-        return new MaxSATSolver(null, MaxSATConfig.builder().build(), Algorithm.INC_WBO);
-    }
-
-    /**
      * Returns a new MaxSAT solver using incremental WBO as algorithm with the MaxSAT configuration from the formula factory.
      * @param f the formula factory
      * @return the MaxSAT solver
@@ -111,33 +101,12 @@ public class MaxSATSolver {
 
     /**
      * Returns a new MaxSAT solver using incremental WBO as algorithm with the given configuration.
-     * @param config the configuration
-     * @return the MaxSAT solver
-     * @deprecated this factory method without formula factory will be removed in the future, use {@link #incWBO(FormulaFactory, MaxSATConfig)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver incWBO(final MaxSATConfig config) {
-        return new MaxSATSolver(null, config, Algorithm.INC_WBO);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using incremental WBO as algorithm with the given configuration.
      * @param f      the formula factory
      * @param config the configuration
      * @return the MaxSAT solver
      */
     public static MaxSATSolver incWBO(final FormulaFactory f, final MaxSATConfig config) {
         return new MaxSATSolver(f, config, Algorithm.INC_WBO);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using LinearSU as algorithm with the default configuration.
-     * @return the MaxSAT solver
-     * @deprecated this factory method without parameter will be removed in the future, use {@link #linearSU(FormulaFactory)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver linearSU() {
-        return new MaxSATSolver(null, MaxSATConfig.builder().cardinality(MaxSATConfig.CardinalityEncoding.MTOTALIZER).build(), Algorithm.LINEAR_SU);
     }
 
     /**
@@ -153,32 +122,11 @@ public class MaxSATSolver {
     /**
      * Returns a new MaxSAT solver using LinearSU as algorithm with the given configuration.
      * @param config the configuration
-     * @return the MaxSAT solver
-     * @deprecated this factory method without formula factory will be removed in the future, use {@link #linearSU(FormulaFactory, MaxSATConfig)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver linearSU(final MaxSATConfig config) {
-        return new MaxSATSolver(null, config, Algorithm.LINEAR_SU);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using LinearSU as algorithm with the given configuration.
-     * @param config the configuration
      * @param f      the formula factory
      * @return the MaxSAT solver
      */
     public static MaxSATSolver linearSU(final FormulaFactory f, final MaxSATConfig config) {
         return new MaxSATSolver(f, config, Algorithm.LINEAR_SU);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using LinearUS as algorithm with the default configuration.
-     * @return the MaxSAT solver
-     * @deprecated this factory method without parameter will be removed in the future, use {@link #linearUS(FormulaFactory)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver linearUS() {
-        return new MaxSATSolver(null, MaxSATConfig.builder().build(), Algorithm.LINEAR_US);
     }
 
     /**
@@ -193,32 +141,11 @@ public class MaxSATSolver {
     /**
      * Returns a new MaxSAT solver using LinearUS as algorithm with the given configuration.
      * @param config the configuration
-     * @return the MaxSAT solver
-     * @deprecated this factory method without formula factory will be removed in the future, use {@link #linearUS(FormulaFactory, MaxSATConfig)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver linearUS(final MaxSATConfig config) {
-        return new MaxSATSolver(null, config, Algorithm.LINEAR_US);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using LinearUS as algorithm with the given configuration.
-     * @param config the configuration
      * @param f      the formula factory
      * @return the MaxSAT solver
      */
     public static MaxSATSolver linearUS(final FormulaFactory f, final MaxSATConfig config) {
         return new MaxSATSolver(f, config, Algorithm.LINEAR_US);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using MSU3 as algorithm with the default configuration.
-     * @return the MaxSAT solver
-     * @deprecated this factory method without parameter will be removed in the future, use {@link #msu3(FormulaFactory)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver msu3() {
-        return new MaxSATSolver(null, MaxSATConfig.builder().build(), Algorithm.MSU3);
     }
 
     /**
@@ -233,32 +160,11 @@ public class MaxSATSolver {
     /**
      * Returns a new MaxSAT solver using MSU3 as algorithm with the given configuration.
      * @param config the configuration
-     * @return the MaxSAT solver
-     * @deprecated this factory method without formula factory will be removed in the future, use {@link #msu3(FormulaFactory, MaxSATConfig)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver msu3(final MaxSATConfig config) {
-        return new MaxSATSolver(null, config, Algorithm.MSU3);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using MSU3 as algorithm with the given configuration.
-     * @param config the configuration
      * @param f      the formula factory
      * @return the MaxSAT solver
      */
     public static MaxSATSolver msu3(final FormulaFactory f, final MaxSATConfig config) {
         return new MaxSATSolver(f, config, Algorithm.MSU3);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using WBO as algorithm with the default configuration.
-     * @return the MaxSAT solver
-     * @deprecated this factory method without parameter will be removed in the future, use {@link #wbo(FormulaFactory)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver wbo() {
-        return new MaxSATSolver(null, MaxSATConfig.builder().build(), Algorithm.WBO);
     }
 
     /**
@@ -273,32 +179,11 @@ public class MaxSATSolver {
     /**
      * Returns a new MaxSAT solver using MSU3 as algorithm with the given configuration.
      * @param config the configuration
-     * @return the MaxSAT solver
-     * @deprecated this factory method without formula factory will be removed in the future, use {@link #wbo(FormulaFactory, MaxSATConfig)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver wbo(final MaxSATConfig config) {
-        return new MaxSATSolver(null, config, Algorithm.WBO);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using MSU3 as algorithm with the given configuration.
-     * @param config the configuration
      * @param f      the formula factory
      * @return the MaxSAT solver
      */
     public static MaxSATSolver wbo(final FormulaFactory f, final MaxSATConfig config) {
         return new MaxSATSolver(f, config, Algorithm.WBO);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using weighted MSU3 as algorithm with the default configuration.
-     * @return the MaxSAT solver
-     * @deprecated this factory method without parameter will be removed in the future, use {@link #wmsu3(FormulaFactory)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver wmsu3() {
-        return new MaxSATSolver(null, MaxSATConfig.builder().incremental(MaxSATConfig.IncrementalStrategy.ITERATIVE).build(), Algorithm.WMSU3);
     }
 
     /**
@@ -309,17 +194,6 @@ public class MaxSATSolver {
     public static MaxSATSolver wmsu3(final FormulaFactory f) {
         final MaxSATConfig conf = new MaxSATConfig((MaxSATConfig) f.configurationFor(ConfigurationType.MAXSAT), MaxSATConfig.IncrementalStrategy.ITERATIVE);
         return new MaxSATSolver(f, conf, Algorithm.WMSU3);
-    }
-
-    /**
-     * Returns a new MaxSAT solver using weighted MSU3 as algorithm with the given configuration.
-     * @param config the configuration
-     * @return the MaxSAT solver
-     * @deprecated this factory method without formula factory will be removed in the future, use {@link #wmsu3(FormulaFactory, MaxSATConfig)} instead
-     */
-    @Deprecated
-    public static MaxSATSolver wmsu3(final MaxSATConfig config) {
-        return new MaxSATSolver(null, config, Algorithm.WMSU3);
     }
 
     /**