+++ src/main/java/org/logicng/transformations/AIGTransformation.java
@@ -44,7 +44,7 @@ import java.util.LinkedHashSet;
 
 /**
  * And-inverter-graph (AIG) transformation.  Returns the AIG of the given formula.
- * @version 2.3.0
+ * @version 3.0.0
  * @since 1.0
  */
 public final class AIGTransformation implements FormulaTransformation {
@@ -53,12 +53,7 @@ public final class AIGTransformation implements FormulaTransformation {
     private boolean cache;
     private static final AIGTransformation INSTANCE = new AIGTransformation();
 
-    /**
-     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
-     * In order to instantiate an object of this class, use the {@link #get()} method.
-     */
-    @Deprecated
-    public AIGTransformation() {
+    private AIGTransformation() {
         // Intentionally left empty
     }