+++ src/main/java/org/logicng/transformations/dnf/DNFSubsumption.java
@@ -44,18 +44,13 @@ import java.util.SortedSet;
  * I.e. performs as many subsumptions as possible.  A subsumption in a DNF means,
  * that e.g. a minterm {@code A & B & C} is subsumed by another minterm {@code A & B}
  * and can therefore be deleted for an equivalent DNF.
- * @version 2.3.0
+ * @version 3.0.0
  * @since 1.5.0
  */
 public final class DNFSubsumption extends Subsumption implements FormulaTransformation {
     private static final DNFSubsumption INSTANCE = new DNFSubsumption();
 
-    /**
-     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
-     * In order to instantiate an object of this class, use the {@link #get()} method.
-     */
-    @Deprecated
-    public DNFSubsumption() {
+    private DNFSubsumption() {
         // Intentionally left empty
     }