+++ src/main/java/org/logicng/transformations/simplification/BackboneSimplifier.java
@@ -42,19 +42,14 @@ import org.logicng.solvers.functions.BackboneFunction;
  * it through the formula. E.g. in the formula {@code A & B & (A | B | C) & (~B | D)}
  * the backbone {@code A, B} is computed and propagated, yielding the simplified formula
  * {@code A & B & D}.
- * @version 2.3.0
+ * @version 3.0.0
  * @since 1.5.0
  */
 public final class BackboneSimplifier implements FormulaTransformation {
 
     private static final BackboneSimplifier INSTANCE = new BackboneSimplifier();
 
-    /**
-     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
-     * In order to instantiate an object of this class, use the {@link #get()} method.
-     */
-    @Deprecated
-    public BackboneSimplifier() {
+    private BackboneSimplifier() {
         // Intentionally left empty
     }