+++ src/main/java/org/logicng/transformations/simplification/DefaultRatingFunction.java
@@ -34,7 +34,7 @@ import org.logicng.formulas.printer.DefaultStringRepresentation;
 /**
  * A rating function which rates a formula with the length of its representation
  * using the {@link DefaultStringRepresentation}.
- * @version 2.3.0
+ * @version 3.0.0
  * @since 2.0.0
  */
 public class DefaultRatingFunction implements RatingFunction<Integer> {
@@ -42,12 +42,7 @@ public class DefaultRatingFunction implements RatingFunction<Integer> {
     private final DefaultStringRepresentation formatter = new DefaultStringRepresentation();
     private static final DefaultRatingFunction INSTANCE = new DefaultRatingFunction();
 
-    /**
-     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
-     * In order to instantiate an object of this class, use the {@link #get()} method.
-     */
-    @Deprecated
-    public DefaultRatingFunction() {
+    private DefaultRatingFunction() {
         // Intentionally left empty
     }