+++ src/main/java/org/logicng/transformations/simplification/DistributiveSimplifier.java
@@ -46,19 +46,14 @@ import java.util.Set;
 
 /**
  * A formula transformation which performs simplifications by applying the distributive laws.
- * @version 2.3.0
+ * @version 3.0.0
  * @since 1.3
  */
 public final class DistributiveSimplifier implements FormulaTransformation {
 
     private static final DistributiveSimplifier INSTANCE = new DistributiveSimplifier();
 
-    /**
-     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
-     * In order to instantiate an object of this class, use the {@link #get()} method.
-     */
-    @Deprecated
-    public DistributiveSimplifier() {
+    private DistributiveSimplifier() {
         // Intentionally left empty
     }