+++ src/main/java/org/logicng/transformations/simplification/NegationSimplifier.java
@@ -46,19 +46,14 @@ import java.util.List;
  * Negation simplifier.
  * <p>
  * Reduces the number of negations for a formula in a greedy manner. The criterion for the simplification is the length of the resulting formula.
- * @version 2.3.0
+ * @version 3.0.0
  * @since 2.0.0
  */
 public final class NegationSimplifier implements FormulaTransformation {
 
     private static final NegationSimplifier INSTANCE = new NegationSimplifier();
 
-    /**
-     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
-     * In order to instantiate an object of this class, use the {@link #get()} method.
-     */
-    @Deprecated
-    public NegationSimplifier() {
+    private NegationSimplifier() {
         // Intentionally left empty
     }